Crate gany

Structs

Package

The data associated with a package

Repository

The data associated with a repository

Functions

add_package

Add a package to your software installation

add_repository

add_repository_with_url

create_package

Package a piece of software for future distribution

drop_package

Drop a package from your software installation

extract_package

Extract a software package

load_repositories

Load repository data from filesystem into memory

push_repository

refresh_remote_packages

Refresh the local package repository with one from a remote software distribution

sync_repositories

upgrade_packages

Upgrade your local packages with newer versions held in a remote software distribution

write_file

Write a file to the filesystem